Lab 5 Part 1 - 17th commit
I made the font colour darker so that the heading can be more easily read and for visual appeal. I made the main text elements easier to read with the sans-serif font, ensured the main heading was large, visually appealing, and consisted of a shadowed font for more contrast, and made the body text a consistent size and line height in the style.css file.