Lab 5 part 1 - 18th commit
Updated the header layout. Made the background colour for the layout containers. Made the content easier to read with improved spacing. Made the nav bar use the flex layout and colour scheme. Updated the nav menu with the lists layout. Made sure the nav items were spaced evenly and the nav links easier to read. The search form is properly aligned with the nav bar. The input fields for the search function are correctly styled for better viewing. This was done on the style.css file.