lab 5 Part 1 - 20th commit
Made the style table with the correct borders and spacing, added the table cells padding and the boarder. Made the data cells aligned in the center. Added a column header background for visual appeal and a stripe effect so it can be more easily read and enjoyed in the style.css file.