lab 5 Part 1 - 22nd commit
Added the styling for the comments section, toggle button, and the subheading in the comments. Made the layout spacing for the comment form and added the pair labels & their inputs with the flex layout. The labels width and alignment and the input spacing and flex are added. Added the submit button style for the forms and the layout for the comment list. The commentors name will now be bolded and made a small change so the comment text takes up more space.